Add files via upload
This script will check Amazon Prime now to find if a specific store has deliveries available or not.  Particularly useful now during the COVID-19 pandemic.

Requires geckodriver and selenium to run.  Modify if you prefer to use Chrome or another browser.  I also have the zip code for 98036; change if you are not in the Seattle area.